FizzBuzz
This is the code that I created using google chrome developer tools. Numbers, Fizz, Buzz and FizzBuzz appear in the console after the code is run, the numbers and word appear until number 100 where the word Buzz appears, after this the next value appears is undefined where the console stops.